+++ core/lib/Thelia/Core/Template/Loop/Area.php
@@ -118,6 +118,7 @@ class Area extends BaseLoop implements PropelSearchLoopInterface
                 ->set('NAME', $area->getName())
                 ->set('POSTAGE', $area->getPostage())
             ;
+	    $this->addOutputFields($loopResultRow, $area);
 
             $loopResult->addRow($loopResultRow);
         }