+++ core/lib/Thelia/Core/Template/Loop/Lang.php
@@ -121,6 +121,7 @@ class Lang extends BaseLoop implements PropelSearchLoopInterface
                 ->set("TIME_FORMAT", $result->getTimeFormat())
                 ->set("POSITION", $result->getPosition())
             ;
+	    $this->addOutputFields($loopResultRow, $result);
 
             $loopResult->addRow($loopResultRow);
         }